Rename src/runtime/auth/server
