+++ src/runtime/composables/useLogout.ts
@@ -1,5 +1,5 @@
 import { useMixpanel } from './useMixpanel'
-import { useLogout as authLogout } from '../auth/server/useLogout'
+import { useLogout as authLogout } from '../auth/useLogout'
 
 export const useLogout = async () => {
   useMixpanel().reset()
