+++ src/runtime/lib/auth/user.ts
@@ -1,2 +1,2 @@
 export { useUser } from '../../composables/useUser'
-export type { TlUser } from '../../auth/shared/types'
+export type { TlUser } from '../../auth/types'
